Bump version, remove circular deps
